refactor: move the KMS adapters into a subsystem-neutral key_custody module
The KMSAdapter protocol, its file/env/HSM backends and the config dispatcher
were named for lineage v2, so every other signing surface read them as
lineage's local abstraction and loaded a PEM off disk instead. They move to
core/security/key_custody.py unchanged; lineage_kms.py re-exports the same
objects, which object-identity matters for because HSM subclass discovery
runs through HSMKMSAdapter.__subclasses__.

Retargets the redundant-handler guard at the file that now holds the code.
